Remove compile warning etc...
- Remove warnings on GCC
- MSVC Compatible
- Adds some Str_replace tests

+++ include/bolib/file/libfile.h
@@ -18,7 +18,7 @@ char* gettmpdir(void);
 
 char* abspath(const char* const);
 
-long fsize(const char* const);
+uint fsize(const char* const);
 
 bool makedir(const char* const);
